+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 ## Cluster Agent Changelog
 
+### 1.8.0
+- Update AppArmor profile to allow netdev creation
+- Update AppArmor and sudoers to allow k3s uninstallation
+
 ### 1.7.4
 - Update AppArmor profile for K3s applications
 
@@ -51,6 +55,12 @@
 
 ## Node Agent Changelog
 
+### 1.8.1
+- Relax system boot detection
+
+### 1.8.0
+- Send no error during initial 5 minutes for agent status
+
 ### 1.7.2
 - Drop TLS on caddy internal endpoint
 
@@ -71,6 +81,9 @@
 - Dependency import updates
 
 ## Platform Observability Agent Changelog
+
+### 1.10.0
+- Fix syslog parsing for health check
 
 ### 1.9.0
 - Add collection of cloud-init service logs